2009-08-25 10:51 UTC+0200 Viktor Szakats (harbour.01 syenar.hu)   * bin/postinst.bat     ! Fixed to work without HB_TOP value. Regression after recent change.
  * contrib/rddsql/sddmy/mysqldd.c
  * contrib/rddsql/sddpg/pgsqldd.c
  * contrib/hbbtree/hb_btree.h
  * contrib/hbbtree/hb_btree.c
  * source/vm/macro.c
  * source/vm/hvm.c
  * source/common/expropt2.c
  * source/compiler/complex.c
    ! Fixed MSVC warnings.
      Please review me, especially the date related ones in core.
    * LONG -> long in hb_compExprReduceTimeStampPut().
    ; NOTE: Number of MSVC (32-bit) -W4 warnings are reduced to 280,
            all of them in RDD code. (plus just few more here and there)

+++ harbour/source/compiler/complex.c
@@ -362,7 +362,7 @@ static BOOL hb_comp_timeDecode( PHB_PP_TOKEN pTime, long * plTime )
    if( !pTime || HB_PP_TOKEN_TYPE( pTime->type ) != HB_PP_TOKEN_RIGHT_CB )
       return FALSE;
 
-   *plTime = ( lHour * 60 + lMinute ) * 60000 + lMilliSec;
+   *plTime = ( long ) ( ( lHour * 60 + lMinute ) * 60000 + lMilliSec );
 
    return TRUE;
 }
@@ -413,7 +413,7 @@ static int hb_comp_dayTimeDecode( PHB_COMP_LEX pLex, PHB_PP_TOKEN pToken,
                                   &iDec, &iWidth ) )
             {
                pDay = pDay->pNext;
-               lDate = hb_dateEncode( lYear, lMonth, lDay );
+               lDate = hb_dateEncode( ( long ) lYear, ( long ) lMonth, ( long ) lDay );
                if( lDate != 0 || ( lYear == 0 && lMonth == 0 && lDay == 0 ) )
                {
                   iType = NUM_DATE;
